move enums and interfaces out of api.ts file
