+++ src/api/data/container/depth-mkt-data-description.ts
@@ -1,0 +1,21 @@
+import { SecType } from "../enum/sec-type";
+
+/**
+ * A container for storing Family Code information.
+ */
+export interface DepthMktDataDescription {
+  /** The exchange code. */
+  exchange?: string;
+
+  /** The security type. */
+  secType?: SecType;
+
+  /** TODO: document */
+  listingExch?: string;
+
+  /** TODO: document */
+  serviceDataType?: string;
+
+  /** TODO: document */
+  aggGroup?: number;
+}
