+++ src/api/data/container/family-code.ts
@@ -1,0 +1,10 @@
+/**
+ * A container for storing Family Code information.
+ */
+export interface FamilyCode {
+  /** The account ID. */
+  accountID?: string;
+
+  /** The family code. */
+  familyCode?: string;
+}
