+++ src/api/data/container/news-provider.ts
@@ -1,0 +1,10 @@
+/**
+ * A container for storing New Provider information.
+ */
+export interface NewsProvider {
+  /** The News Provider code. */
+  providerCode?: string;
+
+  /** The  News Provider name. */
+  providerName?: string;
+}
