+++ src/api/data/container/tag-value.ts
@@ -1,0 +1,10 @@
+/**
+ * Convenience interface to define tag-value pairs
+ */
+export interface TagValue {
+  /** The tag name. */
+  tag: string;
+
+  /** The value. */
+  value: string;
+}
