+++ src/api/data/enum/option-type.ts
@@ -1,0 +1,10 @@
+/**
+ * Option types.
+ */
+export enum OptionType {
+  /** Put option. */
+  Put = "P",
+
+  /** Call option. */
+  Call = "C",
+}
